Move traits/etc into write/input.rs and read/output.rs modules
